refactor: extract recording-json-block as shared component
Move RecordingJsonBlock and its JSON tree helpers (JsonTreeNode,
JsonPrimitiveValue, collectExpandedJsonPaths, isJsonBranch,
getJsonBranchEntries, getJsonBranchSummary) from the traffic recording
admin page into src/components/admin/recording-json-block.tsx so they
can be reused by the logs view in upcoming phases.

Behavior is preserved: the traffic recording page now imports the
shared component and renders identically. Adds component tests
covering helpers, default depth-1 expansion, expand/collapse all,
clipboard copy success/error paths, and branch toggling.

Part of openspec change traffic-recording-logs-integration (phase 1).
