+++ openspec/changes/traffic-recording-logs-integration/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-05-18
